feat: Context-aware platform link
Refactor PlatformRedirectLink into PlatformLink, and allow it to automatically pick up the currently selected platform and generate the appropriately link.
